Initial version of shared_storage that includes signing and hybrid encryption. Definitely not fully tested.  + Lots of refactoring along the StorageLocation classes with StorageFactories    as the new main interface to get StorageMethods
The simple tested example had only one user in the registry the files use:
  signature: 613 bytes for public key and signature
  key: 577 bytes for public key and key-blob
Both also store version information. Sizes can be halfed down since only the user
id should be required, not the public key. Also, a human readable storage would be
appreciated to manually inspect "who has written the data" in the signature and
"who can open it" in the keys.