Tooltip on dropdown to show value.
Two problems:
1) Value not yet properly applied. Root cause is that AppxfSetting
handles it as the base_type is already matching and stores the
input, not the value.
2) Minor but the validation is executed twice. First to check
validity before storing, second upon storing. Two possible solutions:
(a) validation shall try to set and catch the specific Exception.
(b) The validated conversion remembers the last call and output.
(c) Just leave things as they are. Performance beeing a minor issue.